Moving dataset to one place (both used for flex and GPT Neo)
