+++ padthai/flex/__init__.py
@@ -1,2 +1,1 @@
-from .dataset import ListDataset
 from .model import FlexGPT2FewShot
